更新一些依赖
